+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/model/NarratableItemData.kt
@@ -2,7 +2,7 @@ package org.wycliffeassociates.otter.jvm.workbookapp.ui.model
 
 import org.wycliffeassociates.otter.common.data.audio.AudioMarker
 import org.wycliffeassociates.otter.common.data.workbook.Chunk
-import org.wycliffeassociates.otter.common.domain.narration.teleprompter.VerseItemState
+import org.wycliffeassociates.otter.common.domain.narration.statemachine.VerseItemState
 
 class NarratableItemData(
     val chunk: Chunk,
