+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/narration/markers/VerseMarkerControl.kt
@@ -32,9 +32,9 @@ import javafx.scene.layout.Region
 import org.kordamp.ikonli.javafx.FontIcon
 import org.kordamp.ikonli.materialdesign.MaterialDesign
 import org.wycliffeassociates.otter.common.data.audio.AudioMarker
-import org.wycliffeassociates.otter.common.domain.narration.teleprompter.NarrationStateType
-import org.wycliffeassociates.otter.common.domain.narration.teleprompter.VerseItemState
-import org.wycliffeassociates.otter.common.domain.narration.teleprompter.VerseState
+import org.wycliffeassociates.otter.common.domain.narration.statemachine.NarrationStateType
+import org.wycliffeassociates.otter.common.domain.narration.statemachine.VerseItemState
+import org.wycliffeassociates.otter.common.domain.narration.statemachine.VerseState
 import org.wycliffeassociates.otter.jvm.workbookapp.ui.narration.menu.VerseMenu
 import tornadofx.*
 import tornadofx.FX.Companion.messages
